requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
cryptography==1.9
Django==1.11.2
idna==2.5
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.5.2
python-decouple==3.0
pytz==2017.2
requests==2.18.2
six==1.10.0
urllib3==1.22
